+++ src/RainBird/requests/CurrentZoneStateRequest.ts
@@ -1,12 +1,19 @@
 import { Request } from './Request';
 
 export class CurrentZoneStateRequest extends Request {
+  private _page: number;
+
+  constructor(page: number) {
+    super();
+    this._page = page;
+  }
+
   get type(): number {
     return 0x3B;
   }
 
   get page(): number {
-    return 0x00;
+    return this._page;
   }
 
   toBuffer(): Buffer {